feat: Load appbar and menu only when not in live mode
