+++ packages/midi-bricks/src/components/drawer-list/DrawerLoader.jsx
@@ -1,0 +1,3 @@
+import { Drawer } from '@material-ui/core'
+
+export default Drawer
